Refactorization finished
