+++ src/it/skip-parameter/verify.groovy
@@ -1,4 +1,4 @@
-File pseudoThemeDir = new File( basedir, "target/mydoctor-themes/wagon-provider-api" )
+File pseudoThemeDir = new File( basedir, "target/asciidoctor-themes/wagon-provider-api" )
 File dependencyFile = new File( basedir, "target/dependency/junit-4.11.jar" )
 
 assert ! pseudoThemeDir.exists() && ( ! dependencyFile.exists() )
