+++ src/main/resources/META-INF/plexus/components.xml
@@ -3,9 +3,9 @@
         <component>
             <role>org.apache.maven.lifecycle.Lifecycle</role>
             <implementation>org.apache.maven.lifecycle.Lifecycle</implementation>
-            <role-hint>mydoctor-build</role-hint>
+            <role-hint>asciidoctor-lifecycle-build</role-hint>
             <configuration>
-                <id>mydoctor-build</id>
+                <id>asciidoctor-lifecycle-build</id>
                 <phases>
                     <phase>theme-download</phase>
                     <phase>pre-build</phase>
@@ -15,7 +15,7 @@
                     <phase>notify</phase>
                 </phases>
                 <default-phases>
-                    <theme-download>com.coutemeier.maven.plugins:mydoctor-maven-plugin:theme-download</theme-download>
+                    <theme-download>com.coutemeier.maven.plugins:asciidoctor-lifecycle-maven-plugin:theme-download</theme-download>
                     <!-- 
                         This has the inconvenient of generating an extra xml, 
                         because it is invoked the default process-asciidoc goal, outside the executions,
